requirements.txt
httpx==0.23.3
scikit-learn
matplotlib
seaborn
numpy
plotly
openpyxl>=3.0.0
pydantic
requests
OpenAI
Image
pandas
streamlit
Pillow
mediapipe>=0.10.20
opencv-python-headless==4.9.0.80
numpy>=1.21.5
protobuf>=4.25.3
